RanokAI/requirements.txt
fastapi
uvicorn
pyserial
pyttsx3
openai
sounddevice
scipy
